Split custom races and races to different commands. custom races is now only useable by "bot overlord" and "organizes races" roles.
